Don't use `dtype` with `read_csv()`

+++ gt/data/__init__.py
@@ -141,7 +141,7 @@ _illness_dtype = {
     "norm_u": "float64",
 }
 
-countrypops: pd.DataFrame = pd.read_csv(_countrypops_fname, dtype=_countrypops_dtype)  # type: ignore
+countrypops: pd.DataFrame = pd.read_csv(_countrypops_fname)  # type: ignore
 countrypops.__doc__ = """
 Yearly populations of countries from 1960 to 2022
 
@@ -154,7 +154,7 @@ non-existence of the entity during that year.
 """
 
 
-sza: pd.DataFrame = pd.read_csv(_sza_fname, dtype=_sza_dtype)  # type: ignore
+sza: pd.DataFrame = pd.read_csv(_sza_fname)  # type: ignore
 sza.__doc__ = """
 Twice hourly solar zenith angles by month & latitude
 
@@ -166,7 +166,7 @@ are symmetric about noon.
 """
 
 
-gtcars: pd.DataFrame = pd.read_csv(_gtcars_fname, dtype=_gtcars_dtype)  # type: ignore
+gtcars: pd.DataFrame = pd.read_csv(_gtcars_fname)  # type: ignore
 gtcars.__doc__ = """
 Deluxe automobiles from the 2014-2017 period
 
@@ -177,7 +177,7 @@ country of origin for the car manufacturer is also given.
 """
 
 
-sp500: pd.DataFrame = pd.read_csv(_sp500_fname, dtype=_sp500_dtype)  # type: ignore
+sp500: pd.DataFrame = pd.read_csv(_sp500_fname)  # type: ignore
 sp500.__doc__ = """
 Daily S&P 500 Index data from 1950 to 2015
 
@@ -188,7 +188,7 @@ capitalization.
 """
 
 
-pizzaplace: pd.DataFrame = pd.read_csv(_pizzaplace_fname, dtype=_pizzaplace_dtype)  # type: ignore
+pizzaplace: pd.DataFrame = pd.read_csv(_pizzaplace_fname)  # type: ignore
 pizzaplace.__doc__ = """
 A year of pizza sales from a pizza place
 
@@ -204,7 +204,7 @@ Cheese pizza has so many cheeses, I can only offer it in Large Size!').
 """
 
 
-exibble: pd.DataFrame = pd.read_csv(_exibble_fname, dtype=_exibble_dtype)  # type: ignore
+exibble: pd.DataFrame = pd.read_csv(_exibble_fname)  # type: ignore
 exibble.__doc__ = """
 A toy example table for testing with gt: exibble
 
@@ -219,7 +219,7 @@ and two groups (grp_a and grp_b) for experimenting with the `rowname_col` and
 """
 
 
-towny: pd.DataFrame = pd.read_csv(_towny_fname, dtype=_towny_dtype)  # type: ignore
+towny: pd.DataFrame = pd.read_csv(_towny_fname)  # type: ignore
 towny.__doc__ = """
 Populations of all municipalities in Ontario from 1996 to 2021
 
@@ -238,7 +238,7 @@ year and population change values from adjacent census years.
 """
 
 
-metro: pd.DataFrame = pd.read_csv(_metro_fname, dtype=_metro_dtype)  # type: ignore
+metro: pd.DataFrame = pd.read_csv(_metro_fname)  # type: ignore
 metro.__doc__ = """
 The stations of the Paris Metro
 
@@ -262,7 +262,7 @@ Transilien rail network, tramway stations, several major train stations (e.g.,
 Gare du Nord, Gare de l'Est, etc.), and many bus lines.
 """
 
-constants: pd.DataFrame = pd.read_csv(_constants_fname, dtype=_constants_dtype)  # type: ignore
+constants: pd.DataFrame = pd.read_csv(_constants_fname)  # type: ignore
 constants.__doc__ = """
 The fundamental physical constants
 
@@ -278,7 +278,7 @@ Technology (NIST), Gaithersburg, Maryland, USA.
 """
 
 
-illness: pd.DataFrame = pd.read_csv(_illness_fname, dtype=_illness_dtype)  # type: ignore
+illness: pd.DataFrame = pd.read_csv(_illness_fname)  # type: ignore
 illness.__doc__ = """
 Lab tests for one suffering from an illness
